Worked some more on 4.1 to iron out any mistakes

+++ auc-server/src/main/java/app/repositories/BidsRepositoryJpa.java
@@ -62,6 +62,9 @@ public class BidsRepositoryJpa extends AbstractEntityRepositoryJPA<Bid> {
      */
     public Bid deleteById(long id) {
         Bid bid = findById(id);
+
+        if (bid == null) return null;
+
         entityManager.remove(bid);
         return bid;
     }
